Using existing Scroller for all environments except AWS Serverless
