+++ geoportal/src/main/java/com/esri/geoportal/lib/elastic/request/BulkRequest.java
@@ -18,6 +18,7 @@ import com.esri.geoportal.context.AppResponse;
 import com.esri.geoportal.context.GeoportalContext;
 import com.esri.geoportal.lib.elastic.ElasticContext;
 import com.esri.geoportal.lib.elastic.http.ElasticClient;
+import com.esri.geoportal.lib.elastic.http.util.BaseScroller;
 import com.esri.geoportal.lib.elastic.http.util.PITSearchScroller;
 import com.esri.geoportal.lib.elastic.util.AccessUtil;
 import com.esri.geoportal.lib.elastic.util.Scroller;
@@ -145,6 +146,7 @@ public class BulkRequest extends AppRequest {
   
   private AppResponse executeWithHttpClient() throws Exception {
     AppResponse response = new AppResponse();
+    BaseScroller scroller;
     ElasticContext ec = GeoportalContext.getInstance().getElasticContext();
     AccessUtil au = new AccessUtil();
     if (getAdminOnly()) {
@@ -155,11 +157,17 @@ public class BulkRequest extends AppRequest {
     AtomicLong loopCount = new AtomicLong();
     int docsPerRequest = getDocsPerRequest();
     
-    // AWS Serverless - use search_after with PIT, TODO probably remove usage of old scroller
-    
-    PITSearchScroller scroller = newPitSearchScroller(ec);
-    
-   // com.esri.geoportal.lib.elastic.http.util.Scroller scroller = newHttpScroller(ec);
+    // AWS Serverless - use search_after with PIT, 
+    if("serverless".equals(ec.getAwsOpenSearchType()))
+	{
+    	scroller = newPitSearchScroller(ec);
+	}
+    else
+    {
+    	//TODO probably remove usage of old scroller, Elastic Search and Open search behaves differently for point in time so need extensive test
+    	scroller = newHttpScroller(ec);
+    }
+
     scroller.scroll(
       new Consumer<com.esri.geoportal.lib.elastic.http.util.SearchHit>(){
         StringBuilder data = new StringBuilder();
